requirements.txt
torch==2.2.2
nflows==0.14
gym==0.26.2
matplotlib
PyYAML
tensorboard
numpy==1.26.4
